move version to pkg
