+++ tls/handlers/handlers.go
@@ -13,7 +13,7 @@ import (
 	"github.com/jmpsec/osctrl/pkg/queries"
 	"github.com/jmpsec/osctrl/pkg/settings"
 	"github.com/jmpsec/osctrl/pkg/tags"
-	"github.com/jmpsec/osctrl/version"
+	"github.com/jmpsec/osctrl/pkg/version"
 )
 
 const (
